Add automatic versioning to package
Changing and bumping version is old style. There is a bug that if you do
not have access to internet then scm tools kinda fail to provide version
during the run time (not case here), so it is safer to let them write
version to some file, which you then use to present as package version.

This looses any package runtime dependency to scmtools, as they are only
required during the packaging process.

Now you release by tagging (I think Travis-CI is already set for that).